Move progress code to its own module
